Fixed #1893: flyway schema version table creation fails after SQL callback execution
