+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/derby/DerbyDatabase.java
@@ -18,10 +18,9 @@ package org.flywaydb.core.internal.database.derby;
 import org.flywaydb.core.api.configuration.FlywayConfiguration;
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Database;
-import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
 import org.flywaydb.core.internal.database.SqlScript;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -72,18 +71,12 @@ public class DerbyDatabase extends Database {
     }
 
     @Override
-    public SqlScript createSqlScript(String sqlScriptSource) {
-        return new DerbySqlScript(sqlScriptSource);
-    }
-
-    @Override
-    public SqlScript createSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                                     String encoding, boolean mixed
+    protected SqlScript doCreateSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        return new DerbySqlScript(sqlScriptResource, placeholderReplacer, encoding, mixed
+        return new DerbySqlScript(sqlScriptResource, sqlScriptSource, mixed
