+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/oracle/OracleDatabase.java
@@ -18,12 +18,11 @@ package org.flywaydb.core.internal.database.oracle;
 import org.flywaydb.core.api.configuration.FlywayConfiguration;
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Database;
-import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
 import org.flywaydb.core.internal.database.SqlScript;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
+import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
 import org.flywaydb.core.internal.util.StringUtils;
 import org.flywaydb.core.internal.util.jdbc.RowMapper;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -97,18 +96,12 @@ public class OracleDatabase extends Database {
     }
 
     @Override
-    public SqlScript createSqlScript(String sqlScriptSource) {
-        return new OracleSqlScript(sqlScriptSource);
-    }
-
-    @Override
-    public SqlScript createSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                                     String encoding, boolean mixed
+    protected SqlScript doCreateSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        return new OracleSqlScript(sqlScriptResource, placeholderReplacer, encoding, mixed
+        return new OracleSqlScript(sqlScriptResource, sqlScriptSource, mixed
