+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/postgresql/PostgreSQLSqlScript.java
@@ -17,27 +17,21 @@ package org.flywaydb.core.internal.database.postgresql;
 
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Delimiter;
-import org.flywaydb.core.internal.database.SqlScript;
+import org.flywaydb.core.internal.database.ExecutableSqlScript;
 import org.flywaydb.core.internal.database.SqlStatementBuilder;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
 import org.flywaydb.core.internal.util.jdbc.ContextImpl;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 /**
  * PostgreSQL-specific SQL script.
  */
-class PostgreSQLSqlScript extends SqlScript<ContextImpl> {
-    PostgreSQLSqlScript(String sqlScriptSource) {
-        super(sqlScriptSource);
-    }
-
-    PostgreSQLSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                        String encoding, boolean mixed
+class PostgreSQLSqlScript extends ExecutableSqlScript<ContextImpl> {
+    PostgreSQLSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        super(sqlScriptResource, placeholderReplacer, encoding, mixed
+        super(sqlScriptResource, sqlScriptSource, mixed
