+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/redshift/RedshiftDatabase.java
@@ -19,10 +19,9 @@ import org.flywaydb.core.api.configuration.FlywayConfiguration;
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Database;
 import org.flywaydb.core.internal.database.SqlScript;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
 import org.flywaydb.core.internal.util.StringUtils;
 import org.flywaydb.core.internal.util.jdbc.JdbcTemplate;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -83,18 +82,12 @@ public class RedshiftDatabase extends Database<RedshiftConnection> {
     }
 
     @Override
-    public SqlScript createSqlScript(String sqlScriptSource) {
-        return new RedshiftSqlScript(sqlScriptSource);
-    }
-
-    @Override
-    public SqlScript createSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                                     String encoding, boolean mixed
+    protected SqlScript doCreateSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        return new RedshiftSqlScript(sqlScriptResource, placeholderReplacer, encoding, mixed
+        return new RedshiftSqlScript(sqlScriptResource, sqlScriptSource, mixed
