+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/redshift/RedshiftSqlScript.java
@@ -17,27 +17,21 @@ package org.flywaydb.core.internal.database.redshift;
 
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Delimiter;
-import org.flywaydb.core.internal.database.SqlScript;
+import org.flywaydb.core.internal.database.ExecutableSqlScript;
 import org.flywaydb.core.internal.database.SqlStatementBuilder;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
 import org.flywaydb.core.internal.util.jdbc.ContextImpl;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 /**
  * Redshift-specific SQL script.
  */
-class RedshiftSqlScript extends SqlScript<ContextImpl> {
-    RedshiftSqlScript(String sqlScriptSource) {
-        super(sqlScriptSource);
-    }
-
-    RedshiftSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                      String encoding, boolean mixed
+class RedshiftSqlScript extends ExecutableSqlScript<ContextImpl> {
+    RedshiftSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        super(sqlScriptResource, placeholderReplacer, encoding, mixed
+        super(sqlScriptResource, sqlScriptSource, mixed
