+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/sqlite/SQLiteDatabase.java
@@ -18,10 +18,10 @@ package org.flywaydb.core.internal.database.sqlite;
 import org.flywaydb.core.api.configuration.FlywayConfiguration;
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Database;
-import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
+import org.flywaydb.core.internal.database.ExecutableSqlScript;
 import org.flywaydb.core.internal.database.SqlScript;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 import java.sql.Connection;
 import java.sql.Types;
@@ -71,18 +71,12 @@ public class SQLiteDatabase extends Database {
     }
 
     @Override
-    public SqlScript createSqlScript(String sqlScriptSource) {
-        return new SQLiteSqlScript(sqlScriptSource);
-    }
-
-    @Override
-    public SqlScript createSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                                     String encoding, boolean mixed
+    protected SqlScript doCreateSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        return new SQLiteSqlScript(sqlScriptResource, placeholderReplacer, encoding, mixed
+        return new SQLiteSqlScript(sqlScriptResource, sqlScriptSource, mixed
