+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/sqlserver/SQLServerDatabase.java
@@ -19,12 +19,11 @@ import org.flywaydb.core.api.configuration.FlywayConfiguration;
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Database;
 import org.flywaydb.core.internal.database.Delimiter;
+import org.flywaydb.core.internal.database.SqlScript;
 import org.flywaydb.core.internal.exception.FlywayDbUpgradeRequiredException;
 import org.flywaydb.core.internal.exception.FlywaySqlException;
-import org.flywaydb.core.internal.database.SqlScript;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
 import org.flywaydb.core.internal.util.StringUtils;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -122,18 +121,12 @@ public class SQLServerDatabase extends Database {
     }
 
     @Override
-    public SqlScript createSqlScript(String sqlScriptSource) {
-        return new SQLServerSqlScript(sqlScriptSource);
-    }
-
-    @Override
-    public SqlScript createSqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                                     String encoding, boolean mixed
+    protected SqlScript doCreateSqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        return new SQLServerSqlScript(sqlScriptResource, placeholderReplacer, encoding, mixed
+        return new SQLServerSqlScript(sqlScriptResource, sqlScriptSource, mixed
 
 
 
@@ -198,7 +191,7 @@ public class SQLServerDatabase extends Database {
     /**
      * @return Whether this is a SQL Azure database.
      */
-    public boolean isAzure() {
+    boolean isAzure() {
         return azure;
     }
 }
