+++ flyway-core/src/main/java/org/flywaydb/core/internal/database/sybasease/SybaseASESqlScript.java
@@ -17,26 +17,20 @@ package org.flywaydb.core.internal.database.sybasease;
 
 import org.flywaydb.core.api.errorhandler.ErrorHandler;
 import org.flywaydb.core.internal.database.Delimiter;
-import org.flywaydb.core.internal.database.SqlScript;
+import org.flywaydb.core.internal.database.ExecutableSqlScript;
 import org.flywaydb.core.internal.database.SqlStatementBuilder;
-import org.flywaydb.core.internal.util.PlaceholderReplacer;
-import org.flywaydb.core.internal.util.scanner.LoadableResource;
+import org.flywaydb.core.internal.util.scanner.Resource;
 
 /**
  * Sybase ASE-specific SQL script.
  */
-class SybaseASESqlScript extends SqlScript {
-    SybaseASESqlScript(String sqlScriptSource) {
-        super(sqlScriptSource);
-    }
-
-    SybaseASESqlScript(LoadableResource sqlScriptResource, PlaceholderReplacer placeholderReplacer,
-                       String encoding, boolean mixed
+class SybaseASESqlScript extends ExecutableSqlScript {
+    SybaseASESqlScript(Resource sqlScriptResource, String sqlScriptSource, boolean mixed
 
 
 
     ) {
-        super(sqlScriptResource, placeholderReplacer, encoding, mixed
+        super(sqlScriptResource, sqlScriptSource, mixed
